refactor: extract container layout
