+++ frontend/app/layout.tsx
@@ -3,6 +3,7 @@ import "./globals.css";
 import Link from "next/link";
 import Image from "next/image";
 import {PhoneArrowDownLeftIcon} from '@heroicons/react/24/outline';
+import Container from "@/app/ui/Container";
 
 export const metadata: Metadata = {
   title: "Sze Chuan House",
@@ -11,15 +12,15 @@ export const metadata: Metadata = {
 
 export default function RootLayout({
   children,
-}: Readonly<{
+}: {
   children: React.ReactNode;
-}>) {
+}) {
   return (
     <html lang="en">
       <body
         className={`antialiased min-h-[100dvh] grid grid-rows-[auto_1fr_auto]`}
       >
-      <section className={"max-w-372 w-full mx-auto px-10"}>
+      <Container>
           <nav className={"flex justify-between items-center border-b-2 border-white"}>
               <Link href={"/"}>
                   <Image src={"/logo.png"} alt={"Logo of the restaurant"} width={200} height={200}/>
@@ -30,12 +31,12 @@ export default function RootLayout({
                   <li><Link href="/booking" className={"py-3 px-5 text-white border-2 border-white rounded-3xl hover:bg-custom-eggwhite  hover:border-2 hover:border-custom-eggwhite hover:text-black transition-colors duration-300"}>Reserver bord</Link></li>
               </ul>
           </nav>
-      </section>
-          <div>
+      </Container>
+          <section>
             {children}
-          </div>
+          </section>
         <footer className={"bg-custom-red text-white"}>
-            <section className={"max-w-372 w-full mx-auto px-10 flex justify-between px-30 py-8"}>
+            <Container style={"flex justify-between px-30 py-8"}>
                 <section>
                     <h3 className={"text-xl mb-3 uppercase font-bold"}>åpningstider</h3>
                     <ul className={"flex flex-col gap-2"}>
@@ -61,7 +62,7 @@ export default function RootLayout({
                         <a href={"https://www.facebook.com/szechuanhousebergen/"} className={"underline"}>Facebook</a>
                     </address>
                 </section>
-            </section>
+            </Container>
         </footer>
       </body>
     </html>
