+++ frontend/app/page.tsx
@@ -1,6 +1,6 @@
 import Image from "next/image";
 import Link from "next/link";
-import Carousel from "./Carousel";
+import Carousel from "./ui/Carousel";
 
 export default function Home() {
     return (
