+++ frontend/app/ui/Container.tsx
@@ -1,0 +1,6 @@
+export default function Container({children, style = ""} :
+    { children: React.ReactNode; style?: string;}){
+    return <section className={`max-w-372 w-full mx-auto px-10 ${style}`}>
+        {children}
+    </section>
+}
